Add hwservicemanager vintf fragments to system_ext
PRODUCT_HIDL_ENABLED is used to force HIDL to be enabled for device bring-up when
partners have converted all of their AOSP HAL implementations to AIDL,
but still have some partner-owned interfaces that they are working on
converting from HIDL TO AIDL.

To allow for this behavior, we need to use a different manifest fragment
that doesn't have the `max-level="8"` entry on it so hwservicemanager
will see itself as registered.

We have to decouple the manifest fragment from the hwservicemanger binary
so we can choose which one to install based on the PRODUCT_HIDL_ENABLED
variable.

Test: PRODUCT_HIDL_ENABLED = true && adb shell lshal
Test: PRODUCT_HIDL_ENABLED = false && adb shell lshal
Bug: 332548148
Change-Id: I11332f1f09cede729729cc2ed66400769dff43c7